refactor: change dir name
